requirements.txt
# =============================================
# DG Cargo Guardian — 依賴套件清單
# =============================================

# Streamlit UI
streamlit>=1.32.0

# LLM API
openai>=1.30.0

# 環境變數
python-dotenv>=1.0.0

# PDF 解析（未來擴充用）
pdfplumber>=0.10.0

# HTTP 請求（Ollama 備用）
requests>=2.31.0
